backend/requirements.txt
prisma
fastapi
python-dotenv
requests
